orchestration/requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-google==10.14.0
pandas==2.2.0
pyarrow==15.0.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2
dbt-core==1.7.8
dbt-bigquery==1.7.4
